requirements.txt
openai>=1.50.0
langchain==0.3.23
langchain-core==0.3.60
langchain-community==0.3.21
langgraph==0.2.70
langchain-text-splitters==0.3.8
sentence-transformers>=2.7.0
faiss-cpu==1.12.0
numpy>=1.26
pydantic==2.10.6
jinja2
python-dotenv
requests
streamlit